refactor: PieceEntity 레이어 위치 수정 및 PiecesDto 삭제 개선
