+++ src/main/java/chess/repository/entity/PieceEntity.java
@@ -1,23 +1,23 @@
-package chess.dto;
+package chess.repository.entity;
 
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 import chess.repository.mapper.PieceMapper;
 import chess.repository.mapper.PositionMapper;
 
-public class PieceDto {
+public class PieceEntity {
     private final String position;
     private final String team;
     private final String type;
 
-    public PieceDto(String position, String team, String type) {
+    public PieceEntity(String position, String team, String type) {
         this.position = position;
         this.team = team;
         this.type = type;
     }
 
-    public static PieceDto of(Piece piece, Position position) {
-        return new PieceDto(
+    public static PieceEntity of(Piece piece, Position position) {
+        return new PieceEntity(
                 PositionMapper.mapPositionToValue(position),
                 PieceMapper.mapPieceTeamToValue(piece),
                 PieceMapper.mapPieceTypeToValue(piece));
